Adapt to the new design changes in libInterOp.
The new changes internalize the interpreter infrastructure and we do not need
to pass it externally. That simplifies the API and make it more resilient
because often a single change in the implementation might need the interpreter
pointer to be passed externally. In that case that would be a breaking change.